refactor: Move kitchen-sink.ts -> storage-format
